final Client Management code

+++ backend/freelancePM/src/main/java/com/freelance/freelancepm/repository/ClientRepository.java
@@ -4,8 +4,8 @@ import com.freelance.freelancepm.model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-// Interface Segregation: Only exposes CRUD operations for Client
+// Interface Segregation- Only exposes CRUD operations for Client
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long> {
-    // You can add custom query methods if needed
+public interface ClientRepository extends JpaRepository<Client, Integer> {
+
 }
